Rename and move method to requirements file
