+++ CICD/CICD.Common.cs
@@ -276,44 +276,6 @@ public partial class CICD // Common
         return -1;
     }
 
-    private bool NugetPackageDoesNotExist()
-    {
-        nameof(NugetPackageDoesNotExist)
-            .LogRequirementTitle($"Checking that the nuget package does not already exist.");
-
-        var project = this.solution?.GetProject(RepoName);
-        var errors = new List<string>();
-
-        if (project is null)
-        {
-            var exMsg = $"The project named '{ProjectName}' could not be found.";
-            exMsg += $"{Environment.NewLine}Check that the 'ProjectName' param in the parameters.json is set correctly.";
-            throw new Exception(exMsg);
-        }
-
-        var projectVersion = project.GetVersion();
-
-        var nugetService = new NugetDataService();
-
-        var packageVersions = nugetService.GetNugetVersions(RepoName).Result;
-
-        var nugetPackageExists = packageVersions.Any(i => i == projectVersion);
-
-        if (nugetPackageExists)
-        {
-            errors.Add($"The nuget package '{RepoName}' version 'v{projectVersion}' already exists.");
-        }
-
-        if (errors.Count <= 0)
-        {
-            return true;
-        }
-
-        errors.PrintErrors();
-
-        return false;
-    }
-
     private void LogErrorAndFail(string errorMsg, string failMsg)
     {
         Log.Error(errorMsg);
